requirements.txt
asgiref==3.3.1
bokeh==2.2.3
click==7.1.2
cloudpickle==1.6.0
dask==2021.2.0
distributed==2021.2.0
Django==3.1.6
django-filter==2.4.0
djangorestframework==3.12.2
fsspec==0.8.5
HeapDict==1.0.1
Jinja2==2.11.3
locket==0.2.1
MarkupSafe==1.1.1
msgpack==1.0.2
numpy==1.20.1
packaging==20.9
pandas==1.2.2
partd==1.1.0
Pillow==8.1.0
psutil==5.8.0
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
sortedcontainers==2.3.0
SQLAlchemy==1.3.23
sqlparse==0.4.1
tblib==1.7.0
toolz==0.11.1
tornado==6.1
typing-extensions==3.7.4.3
zict==2.0.0
